feat: add Teams client dialog and compose features
