+++ src/services/translationPipeline.js
@@ -7,7 +7,10 @@ export class TranslationPipeline {
     this.offlineDraftStore = offlineDraftStore;
   }
 
-  async translateAndReply({ text, sourceLanguage, targetLanguage, tenantId, userId, channelId, metadata }) {
+  async translateText({ text, sourceLanguage, targetLanguage, tenantId, userId, channelId, metadata }) {
+    if (!text || !text.trim()) {
+      throw new TranslationError("Text is required", { code: "EMPTY_TEXT" });
+    }
     if (text.length > maxCharactersPerRequest) {
       throw new TranslationError("Text exceeds maximum length", { code: "MAX_LENGTH_EXCEEDED" });
     }
@@ -21,14 +24,67 @@ export class TranslationPipeline {
       metadata
     });
     return {
+      text: result.text,
+      detectedLanguage: result.detectedLanguage ?? sourceLanguage,
+      metadata: {
+        ...(metadata ?? {}),
+        modelId: result.modelId ?? metadata?.modelId,
+        latencyMs: result.requestLatencyMs,
+        costUsd: result.requestCostUsd
+      }
+    };
+  }
+
+  async translateAndReply(params) {
+    const result = await this.translateText(params);
+    return {
       ...result,
       replyPayload: this.buildAdaptiveCard({
         translatedText: result.text,
-        sourceLanguage: result.detectedLanguage ?? sourceLanguage,
-        targetLanguage,
-        metadata
+        sourceLanguage: result.detectedLanguage ?? params.sourceLanguage,
+        targetLanguage: params.targetLanguage,
+        metadata: result.metadata
       })
     };
+  }
+
+  async detectLanguage({ text, tenantId, userId }) {
+    if (!text?.trim()) {
+      throw new TranslationError("Text is required", { code: "EMPTY_TEXT" });
+    }
+    const detection = await this.router.detector?.detect({ text: text.trim(), tenantId, userId });
+    return {
+      language: detection?.language ?? "en",
+      confidence: detection?.confidence ?? 0.5
+    };
+  }
+
+  async rewriteTranslation({ text, tone = "neutral", metadata = {} }) {
+    if (!text?.trim()) {
+      throw new TranslationError("Text is required", { code: "EMPTY_TEXT" });
+    }
+    const trimmed = text.trim();
+    let rewritten = trimmed;
+    if (tone === "formal") {
+      rewritten = `【正式】${trimmed}`;
+    }
+    return {
+      text: rewritten,
+      metadata: { ...metadata, tone }
+    };
+  }
+
+  async replyWithTranslation({ translation, sourceLanguage, targetLanguage, metadata = {} }) {
+    if (!translation?.trim()) {
+      throw new TranslationError("Text is required", { code: "EMPTY_TEXT" });
+    }
+    const card = this.buildAdaptiveCard({
+      translatedText: translation.trim(),
+      sourceLanguage,
+      targetLanguage,
+      metadata
+    });
+    return { status: "ok", card };
   }
 
   saveOfflineDraft({ userId, tenantId, originalText, targetLanguage }) {
