+++ src/teamsClient/teamsContext.js
@@ -1,0 +1,1 @@
+export { resolveTeamsSdk, ensureTeamsContext, DEFAULT_CONTEXT } from "../webapp/teamsContext.js";
