+++ src/webapp/compose.html
@@ -19,6 +19,10 @@
       <label class="toggle">
         <input type="checkbox" data-terminology-toggle checked /> 使用术语表
       </label>
+      <label class="toggle">
+        <input type="checkbox" data-tone-toggle /> 正式语气
+      </label>
+      <p class="compose__cost" data-compose-cost></p>
       <label class="compose__editor">
         原文
         <textarea rows="5" data-compose-input placeholder="输入要翻译的消息"></textarea>
