+++ src/webapp/dialog.html
@@ -21,12 +21,16 @@
           源语言
           <select data-source-select aria-label="源语言选择"></select>
         </label>
+        <p class="dialog__detected" data-detected-language></p>
         <label>
           目标语言
           <select data-target-select aria-label="目标语言选择"></select>
         </label>
         <label class="toggle">
           <input type="checkbox" data-terminology-toggle checked /> 使用术语表
+        </label>
+        <label class="toggle">
+          <input type="checkbox" data-tone-toggle /> 正式语气
         </label>
         <p class="dialog__cost" data-cost-hint></p>
       </section>
