+++ src/webapp/settings.html
@@ -21,6 +21,10 @@
         <select data-default-language aria-label="默认目标语言"></select>
       </section>
       <section>
+        <h2>默认语气</h2>
+        <select data-default-tone aria-label="默认语气"></select>
+      </section>
+      <section>
         <label class="toggle">
           <input type="checkbox" data-terminology-toggle checked /> 启用术语库
         </label>
